requirements.txt
anthropic>=0.40.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.0
pytest>=8.3.0
